Refaktorer valideringsmetode ++
