+++ src/overgangsstønad/steg/4-barnasbosted/BarnasBosted.tsx
@@ -14,7 +14,7 @@ import { Stønadstype } from '../../../models/søknad/stønadstyper';
 import { LocationStateSøknad } from '../../../models/søknad/søknad';
 import { logSidevisningOvergangsstonad } from '../../../utils/amplitude';
 import { useMount } from '../../../utils/hooks';
-import { erForelderUtfylt } from '../5-aktivitet/helper';
+import { erForelderUtfylt } from '../../../helpers/steg/forelder';
 
 const scrollTilRef = (ref: RefObject<HTMLDivElement>) => {
   if (!ref || !ref.current) return;
